Reorganized Assistants.withUserData usage
